Zoomable image viewer added
